refactor: broke the application into modules
