+++ utils/globalVariables.js
@@ -1,0 +1,7 @@
+import os from 'os';
+
+export const ROOT_DIR = os.homedir();
+export const dirs = {
+  startDir: ROOT_DIR,
+  curDir: ROOT_DIR
+}
